+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 
 - Serve Files with Restful Endpoint `/files`.
 - Add health check to mutation and query objects.
+- Provide string `"null"` converts to `null` in default value.
 
 ## [v0.0.14]